Undone the parts of [36a89e6461] changing the buffer API. Going back to int's, adapting all users for the same, and putting a compile-time-assert in place to check that its sizeof(int) is good enough (>= 4). Which should be true on most machines this will get compiled on. Small embedded system where this is most likely violated (i.e. int == short) are not a target for crimp anyway.

+++ c/buffer.c
@@ -126,7 +126,7 @@ crimp_buf_read_uint16le (crimp_buffer* buf, unsigned int* res)
 }
 
 void
-crimp_buf_read_uint32le (crimp_buffer* buf, unsigned long* res)
+crimp_buf_read_uint32le (crimp_buffer* buf, unsigned int* res)
 {
     CRIMP_ASSERT_BOUNDS (3,(buf->sentinel - buf->here));
 
@@ -150,7 +150,7 @@ crimp_buf_read_uint16be (crimp_buffer* buf, unsigned int* res)
 }
 
 void
-crimp_buf_read_uint32be (crimp_buffer* buf, unsigned long* res)
+crimp_buf_read_uint32be (crimp_buffer* buf, unsigned int* res)
 {
     CRIMP_ASSERT_BOUNDS (3,(buf->sentinel - buf->here));
